# specs/003-ai-chatbot-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Powered Conversational Todo Chatbot API
  description: API specification for the AI chatbot integration with the Todo application
  version: 1.0.0
  contact:
    name: Todo App Development Team
    email: dev@todoapp.example.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todoapp.example.com
    description: Production server

security:
  - BearerAuth: []

paths:
  /api/{user_id}/chat:
    post:
      summary: Send a message to the AI chatbot
      description: Sends a user message to the AI chatbot and receives a response with potential tool calls
      operationId: sendMessageToChatbot
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
          example: "user_12345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response from the AI chatbot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user trying to access another user's chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Chat

  /api/{user_id}/conversations:
    get:
      summary: Get user's conversation history
      description: Retrieves a list of the user's recent conversations
      operationId: getUserConversations
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
          example: "user_12345"
        - name: limit
          in: query
          required: false
          description: Maximum number of conversations to return
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of conversations to skip for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of user's conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationSummary'
                  total_count:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized - invalid or expired JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user trying to access another user's conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Conversations

  /api/{user_id}/conversations/{conversation_id}:
    get:
      summary: Get a specific conversation
      description: Retrieves all messages in a specific conversation
      operationId: getConversation
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
          example: "user_12345"
        - name: conversation_id
          in: path
          required: true
          description: The ID of the conversation to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Conversation details with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetail'
        '401':
          description: Unauthorized - invalid or expired JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user trying to access another user's conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Conversations

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint

  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The user's message to the chatbot
          example: "Add a task to buy groceries"
        conversation_id:
          type: integer
          description: Optional ID of an existing conversation to continue
          example: 123
        metadata:
          type: object
          description: Additional metadata for the message
          properties:
            timezone:
              type: string
              description: User's timezone
              example: "America/New_York"
            client_info:
              type: object
              description: Information about the client making the request
              properties:
                device:
                  type: string
                  example: "web"
                version:
                  type: string
                  example: "1.0.0"

    ChatResponse:
      type: object
      required:
        - message
        - conversation_id
        - timestamp
      properties:
        message:
          type: string
          description: The AI assistant's response message
          example: "I've added the task 'buy groceries' to your list."
        conversation_id:
          type: integer
          description: The ID of the conversation
          example: 123
        tool_calls:
          type: array
          description: Any tool calls made by the AI agent
          items:
            $ref: '#/components/schemas/ToolCall'
        tool_responses:
          type: array
          description: Responses from tools that were called
          items:
            $ref: '#/components/schemas/ToolResponse'
        confirmation_message:
          type: string
          description: Friendly confirmation of actions taken
          example: "Successfully added task 'buy groceries'"
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
          example: "2023-10-20T10:30:00Z"

    ToolCall:
      type: object
      required:
        - id
        - name
        - arguments
      properties:
        id:
          type: string
          description: Unique identifier for the tool call
          example: "call_abc123"
        name:
          type: string
          description: Name of the tool being called
          enum: [add_task, list_tasks, update_task, complete_task, delete_task]
          example: "add_task"
        arguments:
          type: object
          description: Arguments passed to the tool
          example:
            title: "buy groceries"
            description: "Need to buy milk, bread, and eggs"

    ToolResponse:
      type: object
      required:
        - tool_call_id
        - result
      properties:
        tool_call_id:
          type: string
          description: ID of the tool call this response corresponds to
          example: "call_abc123"
        result:
          type: object
          description: Result of the tool execution
          example:
            success: true
            data:
              id: 456
              title: "buy groceries"
              completed: false
              created_at: "2023-10-20T10:30:00Z"
        error:
          type: object
          description: Error information if the tool call failed
          properties:
            type:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Task title is required"

    ConversationSummary:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the conversation
          example: 123
        title:
          type: string
          description: Title of the conversation
          example: "Grocery Shopping Tasks"
        created_at:
          type: string
          format: date-time
          description: When the conversation was created
          example: "2023-10-20T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the conversation was last updated
          example: "2023-10-20T10:30:00Z"
        message_count:
          type: integer
          description: Number of messages in the conversation
          example: 5

    ConversationDetail:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
        - messages
      properties:
        id:
          type: integer
          description: Unique identifier for the conversation
          example: 123
        title:
          type: string
          description: Title of the conversation
          example: "Grocery Shopping Tasks"
        created_at:
          type: string
          format: date-time
          description: When the conversation was created
          example: "2023-10-20T09:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the conversation was last updated
          example: "2023-10-20T10:30:00Z"
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        participants:
          type: array
          description: List of participants in the conversation
          items:
            type: string
            enum: [user, assistant]
            example: ["user", "assistant"]

    Message:
      type: object
      required:
        - id
        - sender_type
        - content
        - timestamp
        - sequence_number
      properties:
        id:
          type: integer
          description: Unique identifier for the message
          example: 456
        sender_type:
          type: string
          enum: [user, assistant]
          description: Who sent the message
          example: "user"
        content:
          type: string
          description: The message content
          example: "Add a task to buy groceries"
        role:
          type: string
          enum: [user, assistant, tool_call, tool_response]
          description: The role of this message in the conversation context
          example: "user"
        tool_calls:
          type: array
          description: Tool calls made in this message (for assistant messages)
          items:
            $ref: '#/components/schemas/ToolCall'
        tool_responses:
          type: array
          description: Tool responses in this message (for assistant messages)
          items:
            $ref: '#/components/schemas/ToolResponse'
        timestamp:
          type: string
          format: date-time
          description: When the message was created
          example: "2023-10-20T10:30:00Z"
        sequence_number:
          type: integer
          description: Order of the message in the conversation
          example: 3

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "The request was malformed"
        details:
          type: object
          description: Additional error details
          example:
            field: "message"
            reason: "Field is required"

tags:
  - name: Chat
    description: AI chatbot interaction endpoints
  - name: Conversations
    description: Conversation management endpoints